rename Processing to Checkout
