+++ src/App.svelte
@@ -5,14 +5,14 @@
 
   import Footer from "./components/Footer.svelte";
   import Header from "./components/Header.svelte";
-  import Processing from "./components/Processing.svelte";
+  import Checkout from "./components/Checkout.svelte";
   import Starter from "./components/Starter.svelte";
   import Review from "./components/Review.svelte";
 
   let files = [];
   let breadcrumbsPath = [{ href: "/", text: "Home" }];
 
-  const views = [Starter, Processing, Review];
+  const views = [Starter, Checkout, Review];
 
   // breadcrumbs path defined
   const path = [
